requirements.txt
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langfuse>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

